@@ -13,95 +13,143 @@ use App\Controller\UserController;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
 use Symfony\Component\Validator\Constraints as Assert;
 
-/**
- * @ORM\Entity(repositoryClass=UserRepository::class)
- * @ORM\Table(name="`user`")
- * 
- * 
- * @ApiResource(
- *  normalizationContext={"groups"={"user:read"}},
- *  denormalizationContext={"groups"={"user:write"}},
- *  collectionOperations={
- *      "get"={
- *          "pagination_enabled"=false,
- *          "access_control"="is_granted('ROLE_ADMIN')"
- *      },
- *      "post"={
- *          "pagination_enabled"=false
- *      },
- *      "me"={
- *          "method"="GET",
- *          "path"="/users/me",
- *          "controller"=UserController::class,
- *          "pagination_enabled"=false
- *      }
- *  },
- *  itemOperations={
- *      "get" = {
- *          "access_control"="is_granted('ROLE_ADMIN')" 
- *      },
- *      "put" = {
- *          "access_control"="is_granted('ROLE_ADMIN')" 
- *      }, 
- *      "delete" = {
- *          "access_control"="is_granted('ROLE_ADMIN')" 
- *      },
- *      "patch" = {
- *          "access_control"="is_granted('ROLE_ADMIN')" 
- *      }, 
- *  }
- * )
- * 
- * @UniqueEntity(fields={"email"})
- **/
+#[ORM\Entity(repositoryClass: UserRepository::class)]
+#[ORM\Table(name: "`user`")]
+#[ApiResource(
+    normalizationContext: ["groups" => ["user:get"]],
+    denormalizationContext: ["groups" => ["user:post"]],
+    collectionOperations: [
+        "get" => [
+            "pagination_enabled" => false,
+            "access_control" => "is_granted('ROLE_ADMIN')",
+            "normalization_context" => ["groups" => ["user:get"]]
+        ],
+        "post" => [
+            "pagination_enabled" => false,
+            "denormalization_context" => ["groups" => ["user:post"]]
+        ],
+        "me" => [
+            "method" => "GET",
+            "path" => "/users/me",
+            "controller" => UserController::class,
+            "pagination_enabled" => false
+        ]
+    ],
+    validationGroups: ["user:get", "user:post", "user:put"],
+    itemOperations: [
+        "get" => [
+            "normalization_context" => [
+                "groups" => ["user:get"],
+                "access_control" => "is_granted('IS_AUTHENTICATED_FULLY') and object == user  or is_granted('ROLE_ADMIN')"
+            ]
+        ],
+        "put" => [
+            "denormalization_context" => ["groups" => ["user:put"]],
+            "normalization_context" => ["groups" => ["user:get"]],
+            "access_control" => "is_granted('IS_AUTHENTICATED_FULLY') and object == user or is_granted('ROLE_ADMIN')"
+        ],
+        "delete" => [
+            "access_control" => "is_granted('IS_AUTHENTICATED_FULLY') and object == user or is_granted('ROLE_ADMIN')"
+        ],
+        "put-password" => [
+            "method" => "PUT",
+            "path" => "/users/{id}/reset-password",
+            "controller" => ResetPasswordAction::class,
+            "denormalization_context" => ["groups" => ["user:reset:password"]],
+            "access_control" => "is_granted('IS_AUTHENTICATED_FULLY') and object == user"
+        ]
+    ]
+)]
+
+#[ORM\HasLifecycleCallbacks]
+#[UniqueEntity("email")]
 class User implements UserInterface, PasswordAuthenticatedUserInterface
 {
-    /**
-     * @ORM\Id
-     * @ORM\GeneratedValue
-     * @ORM\Column(type="integer")
-     * 
-     * @Groups("user:read")
-     */
+    #[ORM\PrePersist]
+    #[ORM\PreUpdate]
+    public function updatedTimestamps(): void
+    {
+        $this->setUpdatedAt(new \DateTime('now'));
+        if ($this->getCreatedAt() === null) {
+            $this->setCreatedAt(new \DateTime('now'));
+        }
+    }
+
+    #[ORM\Id]
+    #[ORM\GeneratedValue]
+    #[ORM\Column(type: "integer")]
+    #[Groups("user:get")]
     private $id;
 
-    /**
-     * @ORM\Column(type="string", length=180, unique=true)
-     * @Assert\NotBlank()
-     * @Assert\Email()
-     * 
-     * @Groups({"user:read", "user:write"})
-     */
+
+    #[Groups(["user:post", "user:get"])]
+    #[Assert\NotBlank]
+    #[Assert\Email]
+    #[ORM\Column(type: "string", length: 180, unique: true)]
     private $email;
 
-    /**
-     * @ORM\Column(type="json")
-     * 
-     * @Groups("user:read")
-     */
+    #[ORM\Column(type: "json")]
+    #[Groups("user:get")]
     private $roles = [];
 
-    /**
-     * @var string The hashed password
-     * @ORM\Column(type="string")
-     */
+    #[ORM\Column(type: "string")]
+    #[Assert\NotBlank()]
+    #[Assert\Regex(
+        pattern: "/(?=.*[A-Z])(?=.*[a-z])(?=.*[0-9]).{7,}/",
+        message: "Password must be seven characters long and contain at least one digit, one upper case letter and one lower case letter"
+    )]
     private $password;
 
-    /**
-     * @Groups("user:write")
-     * 
-     * @SerializedName("password")
-     * @Assert\NotBlank()
-     */
+    #[SerializedName("password")]
+    #[Assert\NotBlank]
+    #[Groups("user:post")]
     private $plainPassword;
 
-    /**
-     * @ORM\Column(type="string", length=255)
-     * 
-     * @Groups({"user:read", "user:write"})
-     * @Assert\NotBlank()
-     */
+
+    #[ORM\Column(type: "string", length: 255)]
+    #[Groups(["user:post", "user:get"])]
+    #[Assert\NotBlank]
     private $username;
+
+    #[ORM\Column(type: "boolean", options: ["default" => "0"])]
+    private $enabled = false;
+
+    #[ORM\Column(type: "string", length: 40, nullable: true)]
+    private $confirmationToken;
+
+    #[Groups(["user:get"])]
+    #[ORM\Column(type: "datetime")]
+    private $createdAt;
+
+    #[Groups(["user:get"])]
+    #[ORM\Column(type: "datetime", nullable: true)]
+    private $updatedAt;
+
+
+    #[ORM\Column(name: "password_change_date", type: "integer", nullable: true)]
+    private $passwordChangeDate;
+
+    #[Groups(["user:reset:password"])]
+    #[SecurityAssert\UserPassword(message: "Wrong value for your current password")]
+    private $oldPassword;
+
+    #[Groups(["user:reset:password"])]
+    #[Assert\NotBlank(groups: ["user:reset:password"])]
+    #[Assert\Regex(
+        pattern: "/(?=.*[A-Z])(?=.*[a-z])(?=.*[0-9]).{7,}/",
+        message: "Password must be seven characters long and contain at least one digit, one upper case letter and one lower case letter"
+    )]
+    private $newPassword;
+
+
+    #[Groups(["user:reset:password"])]
+    #[Assert\NotBlank(groups: ["user:reset:password"])]
+    #[Assert\Expression(
+        "this.getNewPassword() === this.getNewRetypedPassword()",
+        message: "passwords does not match"
+    )]
+    private $newRetypedPassword;
+
 
     public function getId(): ?int
     {
@@ -204,6 +252,103 @@ class User implements UserInterface, PasswordAuthenticatedUserInterface
     public function setUsername(string $username): self
     {
         $this->username = $username;
+
+        return $this;
+    }
+
+    public function getConfirmationToken(): ?string
+    {
+        return $this->confirmationToken;
+    }
+
+    public function setConfirmationToken(?string $confirmationToken): self
+    {
+        $this->confirmationToken = $confirmationToken;
+
+        return $this;
+    }
+
+    public function getEnabled(): ?bool
+    {
+        return $this->enabled;
+    }
+
+    public function setEnabled(bool $enabled): self
+    {
+        $this->enabled = $enabled;
+
+        return $this;
+    }
+
+    public function getCreatedAt(): ?\DateTimeInterface
+    {
+        return $this->createdAt;
+    }
+
+    public function setCreatedAt(\DateTimeInterface $createdAt): self
+    {
+        $this->createdAt = $createdAt;
+
+        return $this;
+    }
+
+    public function getUpdatedAt(): ?\DateTimeInterface
+    {
+        return $this->updatedAt;
+    }
+
+    public function setUpdatedAt(?\DateTimeInterface $updatedAt): self
+    {
+        $this->updatedAt = $updatedAt;
+
+        return $this;
+    }
+
+
+    public function getPasswordChangeDate()
+    {
+        return $this->passwordChangeDate;
+    }
+
+    public function setPasswordChangeDate($passwordChangeDate)
+    {
+        $this->passwordChangeDate = $passwordChangeDate;
+
+        return $this;
+    }
+
+    public function getNewPassword()
+    {
+        return $this->newPassword;
+    }
+
+    public function setNewPassword($newPassword)
+    {
+        $this->newPassword = $newPassword;
+
+        return $this;
+    }
+
+    public function getNewRetypedPassword()
+    {
+        return $this->newRetypedPassword;
+    }
+
+    public function setNewRetypedPassword($newRetypedPassword)
+    {
+        $this->newRetypedPassword = $newRetypedPassword;
+
+        return $this;
+    }
+
+    public function getOldPassword()
+    {
+        return $this->oldPassword;
+    }
+
+    public function setOldPassword($oldPassword)
+    {
+        $this->oldPassword = $oldPassword;
 
         return $this;
     }
